Even Cleaner Disruption
Further cleaned up disruption so that it's just a set of three
coroutines that activate sequentially instead of an intermediary
updateloop thing that I was doing before.  It just seemed too confusing
and decided that this way is best.

Also updated it so that the blade is "disrupted" only when the blade is
going down.  This should eliminate weird errant hits when you should be
disrupted and hitting a player later on.