Fixed Small Bug With Disruption
Found a small bug with the disruption that took a bit of work to figure
out and fix.  The blade wasn't disrupting properly when activated
already inside another blade.

It still doesn't disrupt 100% properly, but I might have to revamp the
whole disruption process.  I'll cross that bridge when I get to it.